requirements.txt
# Core dependencies
torch>=2.0.0
torchdiffeq>=0.2.3
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Configuration
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Data
huggingface-hub>=0.16.0

# Progress
tqdm>=4.65.0

# Optional: Logging
# wandb>=0.15.0

# Development
pytest>=7.4.0
black>=23.7.0
ruff>=0.0.280
mypy>=1.5.0
